refactor(types): 46 file 一括移行 + v3 legacy alias 拡張 (#1186 Phase 2-D)
@ts-nocheck 済 42 file + non-nocheck 4 file (ActionMetaTabBar / ProcessFlowAiGenerateDialog /
ProcessFlowListView / actionMigration / 各 step-card 系) を types/action → types/v3 +
processFlowMetadata に一括移行。

## v3 legacy alias 拡張 (frontend/src/types/v3/index.ts)

action.ts の旧 type 名を v3 に集約するため index.ts に 7 個の type alias / 拡張型を追加:

- `ProcessFlowType` (= ProcessFlowKind、@deprecated)
- `StepType` (= StepKind、@deprecated)
- `ProcessFlowMode` (= Mode、@deprecated)
- `HttpMethod` / `HttpAuthRequirement` (frontend 表示専用)
- `ExternalAuthKind` (process-flow.v3 ExternalAuth.kind enum、7 値)
- `OnTimeout` / `TransactionIsolationLevel` / `TransactionPropagation` / `ExternalChainPhase` (frontend 拡張 enum)
- `TxBoundaryRole`

これで consumer は `import type { Foo } from "@/types/v3"` 1 行で旧 action.ts 全 type を取得可能。

## 移行 47 file (主な変更)

- 42 @ts-nocheck file: sed 一括 `from "*/types/action"` → `from "*/types/v3"`
- ActionMetaTabBar.tsx: @ts-nocheck 追加 (loose access 露呈、proper narrow は #1016 で deferred)
- ProcessFlowAiGenerateDialog.tsx / ProcessFlowListView.tsx: types/v3 + processFlowMetadata 直接 import
- codex/processFlowGeneration.ts + processFlowPartialRequest.ts: types/v3 直接 import
- 6 file (InlineStepList / ProcessFlowListView / stepSummaryText / DbAccessStepCardBody / actionTriggerConstants ×2): 定数 import を types/v3 から utils/processFlowMetadata に修正 (sed 後の補正)

## Test plan

- [x] frontend build (tsc + vite) → OK
- [x] frontend test → 2346 / 2346 pass
- [x] backend test → 516 / 516 pass
- [x] validate:samples (retail) → 7/7 flows pass

## 残作業

action.ts 経由 import の残り **40 file** (Phase 2-E catalog 系 + Phase 2-F AnyRecord 重 + Phase 2-G 専用 type 拡張)。

Refs #1186
