+++ frontend/src/components/process-flow/internal/step-cards/CommonProcessStepCardBody.tsx
@@ -1,7 +1,7 @@
 // @ts-nocheck -- StepCard と同じ legacy/v3 union 緩和理由 (#1016)
 // Phase-2 (#1145): StepCard.tsx の `step.kind === "commonProcess"` body を抽出。
 
-import type { Step } from "../../../../types/action";
+import type { Step } from "../../../../types/v3";
 import type {
   StepCardBodyBaseProps,
   StepCardBodyCommonGroupsProps,
